+++ app/src/main/java/com/example/personality_style_test/bankingtest/BankingTest2.java
@@ -1,6 +1,5 @@
 package com.example.personality_style_test.bankingtest;
 
-import android.content.Intent;
 import android.os.Bundle;
 
 import androidx.fragment.app.Fragment;
@@ -15,21 +14,22 @@ import android.widget.Button;
 import com.example.personality_style_test.R;
 
 
-public class BankingTest_4 extends Fragment {
+public class BankingTest2 extends Fragment {
+
 
     //결과값 이동
     private String bankingTestResult1;
     private String bankingTestResult2;
     private String bankingTestResult3;
-
-    public BankingTest_4() {
+    public BankingTest2() {
         // Required empty public constructor
     }
 
 
-    public static BankingTest_4 newInstance() {
-        BankingTest_4 fragment = new BankingTest_4();
+    public static BankingTest2 newInstance() {
+        BankingTest2 fragment = new BankingTest2();
         Bundle args = new Bundle();
+
         return fragment;
     }
 
@@ -43,26 +43,24 @@ public class BankingTest_4 extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
 
-        //뷰 선언
-        View BankingTest4 = inflater.inflate(R.layout.fragment_banking_test_4, container, false);
+        View BankingTest2 = inflater.inflate(R.layout.fragment_banking_test_2, container, false);
 
-        //버튼 연결
-        Button bankingTest4_an1 = (Button) BankingTest4.findViewById(R.id.bankingtest_4_an1);
-        Button bankingTest4_an2 = (Button) BankingTest4.findViewById(R.id.bankingtest_4_an2);
-        Button bankingTest4_an3 = (Button) BankingTest4.findViewById(R.id.bankingtest_4_an3);
+        Button bankingTest2_an1 = (Button) BankingTest2.findViewById(R.id.bankingtest_2_an1);
+        Button bankingTest2_an2 = (Button) BankingTest2.findViewById(R.id.bankingtest_2_an2);
+        Button bankingTest2_an3 = (Button) BankingTest2.findViewById(R.id.bankingtest_2_an3);
 
         //전달한 번들 메소드 받기
         Bundle bundle = getArguments();
 
         //값이 있다면
-        if(bundle != null) {
+        if(bundle != null){
             bankingTestResult1 = bundle.getString("bankingTestResult1");
             bankingTestResult2 = bundle.getString("bankingTestResult2");
             bankingTestResult3 = bundle.getString("bankingTestResult3");
 
 
             //데이터 전달 및 프래그먼트 변경
-            BankingTest_5 bankingTest_5 = new BankingTest_5();
+            BankingTest3 BankingTest3 = new BankingTest3();
             //전달할 번들 생성
             Bundle bundleNext = new Bundle();
 
@@ -70,8 +68,8 @@ public class BankingTest_4 extends Fragment {
             FragmentManager mnger = getActivity().getSupportFragmentManager();
             FragmentTransaction transaction = mnger.beginTransaction();
 
-
-            bankingTest4_an1.setOnClickListener(new View.OnClickListener() {
+            //버튼 1번 클릭
+            bankingTest2_an1.setOnClickListener(new View.OnClickListener() {
                 @Override
                 public void onClick(View view) {
 
@@ -80,16 +78,19 @@ public class BankingTest_4 extends Fragment {
                     bundleNext.putString("bankingTestResult2",bankingTestResult2);
                     bundleNext.putString("bankingTestResult3",bankingTestResult3);
                     //다음 프래그먼트로 값 이동
-                    bankingTest_5.setArguments(bundleNext);
+                    BankingTest3.setArguments(bundleNext);
 
                     //이동
-                    transaction.replace(R.id.bankingtest_start_fragment, bankingTest_5);
+                    transaction.replace(R.id.bankingtest_start_fragment, BankingTest3);
 
                     transaction.commit();
+
+
                 }
             });
 
-            bankingTest4_an2.setOnClickListener(new View.OnClickListener() {
+            //2번 버튼 클릭
+            bankingTest2_an2.setOnClickListener(new View.OnClickListener() {
                 @Override
                 public void onClick(View view) {
 
@@ -97,17 +98,17 @@ public class BankingTest_4 extends Fragment {
                     bundleNext.putString("bankingTestResult1",bankingTestResult1);
                     bundleNext.putString("bankingTestResult2",bankingTestResult2);
                     bundleNext.putString("bankingTestResult3",bankingTestResult3);
-                    //다음 프래그먼트로 값 이동
-                    bankingTest_5.setArguments(bundleNext);
+                    BankingTest3.setArguments(bundleNext);
 
                     //이동
-                    transaction.replace(R.id.bankingtest_start_fragment, bankingTest_5);
+                    transaction.replace(R.id.bankingtest_start_fragment, BankingTest3);
 
                     transaction.commit();
                 }
             });
 
-            bankingTest4_an3.setOnClickListener(new View.OnClickListener() {
+            //3번 버튼 클릭
+            bankingTest2_an3.setOnClickListener(new View.OnClickListener() {
                 @Override
                 public void onClick(View view) {
 
@@ -115,17 +116,22 @@ public class BankingTest_4 extends Fragment {
                     bundleNext.putString("bankingTestResult1",bankingTestResult1);
                     bundleNext.putString("bankingTestResult2",bankingTestResult2);
                     bundleNext.putString("bankingTestResult3",bankingTestResult3);
-                    //다음 프래그먼트로 값 이동
-                    bankingTest_5.setArguments(bundleNext);
+                    BankingTest3.setArguments(bundleNext);
 
                     //이동
-                    transaction.replace(R.id.bankingtest_start_fragment, bankingTest_5);
+                    transaction.replace(R.id.bankingtest_start_fragment, BankingTest3);
 
                     transaction.commit();
                 }
             });
+
+
         }
 
-        return BankingTest4;
+
+
+
+        return BankingTest2;
     }
+
 }
