+++ app/src/main/java/com/example/personality_style_test/bankingtest/BankingTest3.java
@@ -1,6 +1,5 @@
 package com.example.personality_style_test.bankingtest;
 
-import android.content.Intent;
 import android.os.Bundle;
 
 import androidx.fragment.app.Fragment;
@@ -15,21 +14,21 @@ import android.widget.Button;
 import com.example.personality_style_test.R;
 
 
-public class BankingTest_5 extends Fragment {
+public class BankingTest3 extends Fragment {
 
+    //결과값 이동
     private String bankingTestResult1;
     private String bankingTestResult2;
     private String bankingTestResult3;
 
-    public BankingTest_5() {
-
+    public BankingTest3() {
+        // Required empty public constructor
     }
 
-
-    public static BankingTest_5 newInstance() {
-        BankingTest_5 fragment = new BankingTest_5();
+    public static BankingTest3 newInstance() {
+        BankingTest3 fragment = new BankingTest3();
         Bundle args = new Bundle();
-        fragment.setArguments(args);
+
         return fragment;
     }
 
@@ -44,12 +43,11 @@ public class BankingTest_5 extends Fragment {
                              Bundle savedInstanceState) {
 
         //뷰 선언
-        View BankingTest5 = inflater.inflate(R.layout.fragment_banking_test_5, container, false);
+        View BankingTest3 = inflater.inflate(R.layout.fragment_banking_test_3, container, false);
 
-        Button bankingTest5_an1 = (Button) BankingTest5.findViewById(R.id.bankingtest_5_an1);
-        Button bankingTest5_an2 = (Button) BankingTest5.findViewById(R.id.bankingtest_5_an2);
-        Button bankingTest5_an3 = (Button) BankingTest5.findViewById(R.id.bankingtest_5_an3);
-        Button bankingTest5_an4 = (Button) BankingTest5.findViewById(R.id.bankingtest_5_an3);
+        Button bankingTest3_an1 = (Button) BankingTest3.findViewById(R.id.bankingtest_3_an1);
+        Button bankingTest3_an2 = (Button) BankingTest3.findViewById(R.id.bankingtest_3_an2);
+        Button bankingTest3_an3 = (Button) BankingTest3.findViewById(R.id.bankingtest_3_an3);
 
         //전달한 번들 메소드 받기
         Bundle bundle = getArguments();
@@ -62,7 +60,7 @@ public class BankingTest_5 extends Fragment {
 
 
             //데이터 전달 및 프래그먼트 변경
-            BankingTest_6 bankingTest_6 = new BankingTest_6();
+            BankingTest4 bankingTest_4 = new BankingTest4();
             //전달할 번들 생성
             Bundle bundleNext = new Bundle();
 
@@ -70,8 +68,7 @@ public class BankingTest_5 extends Fragment {
             FragmentManager mnger = getActivity().getSupportFragmentManager();
             FragmentTransaction transaction = mnger.beginTransaction();
 
-
-            bankingTest5_an1.setOnClickListener(new View.OnClickListener() {
+            bankingTest3_an1.setOnClickListener(new View.OnClickListener() {
                 @Override
                 public void onClick(View view) {
 
@@ -80,16 +77,16 @@ public class BankingTest_5 extends Fragment {
                     bundleNext.putString("bankingTestResult2",bankingTestResult2);
                     bundleNext.putString("bankingTestResult3",bankingTestResult3);
                     //다음 프래그먼트로 값 이동
-                    bankingTest_6.setArguments(bundleNext);
+                    bankingTest_4.setArguments(bundleNext);
 
                     //이동
-                    transaction.replace(R.id.bankingtest_start_fragment, bankingTest_6);
+                    transaction.replace(R.id.bankingtest_start_fragment, bankingTest_4);
 
                     transaction.commit();
                 }
             });
 
-            bankingTest5_an2.setOnClickListener(new View.OnClickListener() {
+            bankingTest3_an2.setOnClickListener(new View.OnClickListener() {
                 @Override
                 public void onClick(View view) {
 
@@ -98,52 +95,35 @@ public class BankingTest_5 extends Fragment {
                     bundleNext.putString("bankingTestResult2",bankingTestResult2);
                     bundleNext.putString("bankingTestResult3",bankingTestResult3);
                     //다음 프래그먼트로 값 이동
-                    bankingTest_6.setArguments(bundleNext);
+                    bankingTest_4.setArguments(bundleNext);
 
                     //이동
-                    transaction.replace(R.id.bankingtest_start_fragment, bankingTest_6);
+                    transaction.replace(R.id.bankingtest_start_fragment, bankingTest_4);
 
                     transaction.commit();
                 }
             });
 
-
-            bankingTest5_an3.setOnClickListener(new View.OnClickListener() {
+            bankingTest3_an3.setOnClickListener(new View.OnClickListener() {
                 @Override
                 public void onClick(View view) {
 
+                    //해당 값이 null이라도 이 버튼을 누른다면 다음 프래그먼트때는 해당 값을 넘겨야한다.
+                    bankingTestResult3 = "result3";
                     //번틀 값 설정
                     bundleNext.putString("bankingTestResult1",bankingTestResult1);
                     bundleNext.putString("bankingTestResult2",bankingTestResult2);
                     bundleNext.putString("bankingTestResult3",bankingTestResult3);
                     //다음 프래그먼트로 값 이동
-                    bankingTest_6.setArguments(bundleNext);
+                    bankingTest_4.setArguments(bundleNext);
 
                     //이동
-                    transaction.replace(R.id.bankingtest_start_fragment, bankingTest_6);
-
-                    transaction.commit();
-                }
-            });
-
-            bankingTest5_an4.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View view) {
-
-                    //번틀 값 설정
-                    bundleNext.putString("bankingTestResult1",bankingTestResult1);
-                    bundleNext.putString("bankingTestResult2",bankingTestResult2);
-                    bundleNext.putString("bankingTestResult3",bankingTestResult3);
-                    //다음 프래그먼트로 값 이동
-                    bankingTest_6.setArguments(bundleNext);
-
-                    //이동
-                    transaction.replace(R.id.bankingtest_start_fragment, bankingTest_6);
+                    transaction.replace(R.id.bankingtest_start_fragment, bankingTest_4);
 
                     transaction.commit();
                 }
             });
         }
-        return BankingTest5;
+        return BankingTest3;
     }
 }
