+++ app/src/main/res/layout/activity_bankingtest.xml
@@ -6,7 +6,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".bankingtest.bankingtestActivity">
+    tools:context=".bankingtest.BankingTestActivity">
 
     <FrameLayout
         android:layout_width="match_parent"
