+++ app/src/main/res/layout/fragment_banking_test_2.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".bankingtest.BankingTest_2">
+    tools:context=".bankingtest.BankingTest2">
 
     <ImageView
         android:id="@+id/imageView4"
