+++ app/src/main/res/layout/fragment_banking_test_3.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".bankingtest.BankingTest_3">
+    tools:context=".bankingtest.BankingTest3">
 
     <ImageView
         android:id="@+id/imageView5"
