+++ app/src/main/res/layout/fragment_banking_test_5.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".bankingtest.BankingTest_5">
+    tools:context=".bankingtest.BankingTest5">
 
 
     <ImageView
