+++ app/src/main/res/layout/fragment_banking_test_6.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".bankingtest.BankingTest_6">
+    tools:context=".bankingtest.BankingTest6">
 
 
     <ImageView
